DEV: databroker api cleanup, move config to 00-startup
